requirements.txt
ccxt==4.0.66
Django==4.2.4
okx==1.0.8
openai==0.27.8
openpyxl==3.1.2
pandas==2.0.3
pytz==2023.3
Requests==2.31.0
SocksiPy_branch==1.01
websocket_client==1.6.1
django_extensions==3.2.3
mysqlclient==2.2.0
